+++ src/state/github.ts
@@ -1,18 +1,13 @@
-import Octokit, {
-  PullsGetResponse,
-  PullsListResponseItem,
-  ReposGetResponse
-} from "@octokit/rest";
+import Octokit, { ReposGetResponse } from "@octokit/rest";
 import { computed, observable } from "mobx";
-import { loadPullRequest, loadPullRequests } from "../github/api/pull-requests";
-import { loadRepos, ReposListResponse } from "../github/api/repos";
-import { loadReviews, PullsListReviewsResponse } from "../github/api/reviews";
+import { loadRepos } from "../github/api/repos";
 import {
   GetAuthenticatedUserResponse,
   loadAuthenticatedUser
 } from "../github/api/user";
-import { repoWasPushedAfter } from "./filtering/repos-pushed-after";
 import { isReviewNeeded } from "./filtering/review-needed";
+import { refreshOpenPullRequests } from "./loading/pull-requests";
+import { loadAllReviews } from "./loading/reviews";
 import { lastErrorStorage } from "./storage/error";
 import {
   LastCheck,
@@ -73,34 +68,10 @@ export class GitHubState {
       this.repoList,
       this.lastCheck
     );
-    const reviews = await Promise.all(
-      openPullRequests.map(
-        async (
-          pr
-        ): Promise<[string /* node ID */, PullsListReviewsResponse]> => [
-          pr.node_id,
-          await loadReviews(
-            octokit,
-            pr.base.repo.owner.login,
-            pr.base.repo.name,
-            pr.number
-          )
-        ]
-      )
+    const reviewsPerPullRequest = await loadAllReviews(
+      octokit,
+      openPullRequests
     );
-    const reviewsPerPullRequest = reviews.reduce<{
-      [pullRequestNodeId: string]: PullsListReviewsResponse;
-    }>((acc, [nodeId, reviews]) => {
-      acc[nodeId] = reviews;
-      return acc;
-    }, {});
-
-    // For each pull request, check:
-    // - author (is current user or not)
-    // - date of last review by current user
-    // - date of last update
-    //
-    // Then use that information at display time to know what to show.
     this.lastCheck = {
       openPullRequests: openPullRequests.map(pr =>
         pullRequestFromResponse(pr, reviewsPerPullRequest[pr.node_id])
@@ -143,56 +114,4 @@ export class GitHubState {
     }
     this.status = "loaded";
   }
-}
-
-/**
- * Refreshes the list of pull requests for a list of repositories, optimizing
- * for the minimum number of API requests to GitHub as brute-forcing would
- * quickly go over API rate limits if the user has several hundred repositories
- * or many pull requests opened.
- */
-export async function refreshOpenPullRequests(
-  octokit: Octokit,
-  repoList: ReposListResponse,
-  lastCheck: LastCheck | null
-): Promise<Array<PullsListResponseItem | PullsGetResponse>> {
-  // Look for new pull requests in repos that have been recently pushed.
-  const reposWithPotentiallyNewPullRequests = repoList.filter(
-    repoWasPushedAfter(lastCheck ? lastCheck.maximumPushedAt : null)
-  );
-
-  // For each recently pushed repo, load all open pull requests.
-  const openPullRequests: Array<
-    PullsListResponseItem | PullsGetResponse
-  > = (await Promise.all(
-    reposWithPotentiallyNewPullRequests.map(repo =>
-      loadPullRequests(octokit, repo.owner.login, repo.name, "open")
-    )
-  )).flat();
-
-  // Make sure not to do redundant work in the upcoming loop.
-  const alreadyLoadedPullRequestNodeIds = new Set(
-    openPullRequests.map(pr => pr.node_id)
-  );
-
-  // Also update the status of every other known pull request.
-  if (lastCheck) {
-    const updatedPullRequests = await Promise.all(
-      lastCheck.openPullRequests
-        .filter(pr => !alreadyLoadedPullRequestNodeIds.has(pr.nodeId))
-        .map(pr =>
-          loadPullRequest(
-            octokit,
-            pr.repoOwner,
-            pr.repoName,
-            pr.pullRequestNumber
-          )
-        )
-    );
-    openPullRequests.push(
-      ...updatedPullRequests.filter(pr => pr.state === "open")
-    );
-  }
-
-  return openPullRequests;
 }
